Be feat/lookbook

See merge request s10-webmobile1-sub2/S10P12E203!91

Former-commit-id: 0838958901a59de4b9fb34697068333b9bd0b1f5

+++ gawm_web_server/gawm/src/main/java/com/cute/gawm/domain/lookbook/service/LookbookService.java
@@ -1,9 +1,6 @@
 package com.cute.gawm.domain.lookbook.service;
 
-import com.cute.gawm.common.exception.ClothesNotFoundException;
-import com.cute.gawm.common.exception.TagNotFoundException;
-import com.cute.gawm.common.exception.UserNotFoundException;
-import com.cute.gawm.common.exception.UserNotMatchException;
+import com.cute.gawm.common.exception.*;
 import com.cute.gawm.common.response.PagingResponse;
 import com.cute.gawm.common.util.s3.S3Uploader;
 import com.cute.gawm.domain.bookmark.entity.Bookmark;
@@ -13,10 +10,12 @@ import com.cute.gawm.domain.clothes.entity.Clothes;
 import com.cute.gawm.domain.clothes.repository.ClothesRepository;
 import com.cute.gawm.domain.clothes_lookbook.entity.ClothesLookbook;
 import com.cute.gawm.domain.clothes_lookbook.repository.ClothesLookbookRepository;
+import com.cute.gawm.domain.comment.dto.response.CommentResponse;
 import com.cute.gawm.domain.comment.entity.Comment;
 import com.cute.gawm.domain.comment.repository.CommentRepository;
 import com.cute.gawm.domain.following.entity.Following;
 import com.cute.gawm.domain.following.repository.FollowingRepository;
+import com.cute.gawm.domain.following.service.FollowService;
 import com.cute.gawm.domain.like.entity.Likes;
 import com.cute.gawm.domain.like.repository.LikesRepository;
 import com.cute.gawm.domain.lookbook.dto.request.LookbookCreateRequest;
@@ -28,6 +27,7 @@ import com.cute.gawm.domain.lookbook.entity.Lookbook;
 import com.cute.gawm.domain.lookbook.repository.LookbookRepository;
 import com.cute.gawm.domain.lookbook_image.entity.LookbookImage;
 import com.cute.gawm.domain.lookbook_image.repository.LookbookImageRepository;
+import com.cute.gawm.domain.tag.dto.response.TagResponse;
 import com.cute.gawm.domain.tag.entity.Tag;
 import com.cute.gawm.domain.tag.repository.TagRepository;
 import com.cute.gawm.domain.tag_lookbook.entity.TagLookbook;
@@ -65,19 +65,25 @@ public class LookbookService {
     private final BookmarkRepository bookmarkRepository;
     private final FollowingRepository followingRepository;
     private final LikesRepository likesRepository;
+    private final FollowService followService;
 
-    public LookbookResponse getLookbook(final int lookbookId) {
+    public LookbookResponse getLookbook(final int sessionUserId,final int lookbookId) {
+        User sessionUser = userRepository.findById(sessionUserId).orElseThrow(() -> new UserNotFoundException("해당 유저가 존재하지 않습니다."));
+
         final Lookbook lookbook = lookbookRepository.findByLookbookId(lookbookId);
+        if(lookbook==null) throw new DataNotFoundException("해당 룩북은 존재하지 않습니다.");
         final List<ClothesLookbook> clotheLookbooks = clothesLookbookRepository.getAllByLookbookId(lookbookId);
         final List<TagLookbook> tagLookbooks = tagLookbookRepository.getAllByLookbookId(lookbookId);
-        final List<Comment> comments = commentRepository.getAllByLookbookId(lookbookId);
-
+        final List<Comment> commentList = commentRepository.getAllByLookbookId(lookbookId);
+        List<LookbookImage> lookbookImages = lookbookImageRepository.findAllByLookbook_LookbookId(lookbookId);
+        final User user=lookbook.getUser();
 
         List<ClothesMiniResponse> miniResponses = new ArrayList<>();
         clotheLookbooks.forEach(clotheLookbook -> {
             final Clothes clothes = clotheLookbook.getClothes();
             final int id = clothes.getClothesId();
             ClothesMiniResponse clotheMiniResp = ClothesMiniResponse.builder()
+
                     .clothesId(id)
                     .name(clothes.getName())
                     .brand(clothes.getBrand())
@@ -86,18 +92,42 @@ public class LookbookService {
             miniResponses.add(clotheMiniResp);
         });
 
-        List<Tag> tags = new ArrayList<>();
-        tagLookbooks.forEach(tag -> {
-            tags.add(tag.getTag());
+        List<String> lookbookImgs=lookbookImages.stream().map(LookbookImage->LookbookImage.getImage()).collect(Collectors.toList());
+        Integer likeCnt = likesRepository.countByLookbook(lookbook);
+
+        List<TagResponse> tags=tagLookbooks.stream().map(tagLookbook -> new TagResponse(tagLookbook.getTag())).collect(Collectors.toList());
+
+        List<CommentResponse> comments=new ArrayList<>();
+        commentList.forEach(comment -> {
+            CommentResponse commentResp = CommentResponse.builder()
+                    .commentId(comment.getCommentId())
+                    .content(comment.getContent())
+                    .userNickname(comment.getUser().getNickname())
+                    .userProfileImg(comment.getUser().getProfileImg())
+                    .isCommentAuthor(comment.getUser().equals(sessionUser))
+                    .build();
+            comments.add(commentResp);
         });
 
+        boolean isLiked=likesRepository.existsByLookbookAndUserUserId(lookbook,sessionUserId);
+        boolean isBookmarked=bookmarkRepository.existsByLookbookAndUserUserId(lookbook,sessionUserId);
+        boolean isFollowed=followService.isFollowing(sessionUserId,user.getUserId());
+
         return LookbookResponse.builder()
-                .userId(lookbook.getUser().getUserId())
+                .lookbookId(lookbookId)
+                .userId(user.getUserId())
+                .userNickname(user.getNickname())
+                .userProfileImg(user.getProfileImg())
                 .createdAt(lookbook.getCreatedAt())
                 .clothes(miniResponses)
+                .lookbookImgs(lookbookImgs)
+                .likeCnt(likeCnt)
                 .view(lookbook.getView())
                 .tag(tags)
                 .comment(comments)
+                .isLiked(isLiked)
+                .isBookmarked(isBookmarked)
+                .isFollowed(isFollowed)
                 .build();
     }
 
@@ -131,11 +161,13 @@ public class LookbookService {
         });
 
         lookbookRequest.getTags().forEach(tagName -> {
-            Tag tag = Tag.builder()
-                    .name(tagName)
-                    .build();
-            tagRepository.save(tag);
-
+            Tag tag=tagRepository.findByName(tagName);
+            if(tag==null){
+                tag = Tag.builder()
+                        .name(tagName)
+                        .build();
+                tagRepository.save(tag);
+            }
             TagLookbook tagLookbook = TagLookbook.builder()
                     .lookbook(lookbook)
                     .tag(tag)
@@ -144,23 +176,24 @@ public class LookbookService {
         });
     }
 
-    public PagingResponse<List<LookbookMiniResponse>> getLookbooks(Pageable pageable) {
+    public PagingResponse<List<LookbookThumbnailResponse>> getLookbooks(Pageable pageable) {
         Page<Lookbook> lookbooks = lookbookRepository.findAllLookbook(pageable);
-        List<LookbookMiniResponse> lookbookResponse = new ArrayList<>();
+        List<LookbookThumbnailResponse> lookbookResponse = new ArrayList<>();
 
         lookbooks.forEach(lookbook -> {
-            List<LookbookImage> lookbookImages = lookbookImageRepository.findAllByLookbook_LookbookId(lookbook.getLookbookId());
-            List<String> images = new ArrayList<>();
-
-
-            LookbookMiniResponse response = LookbookMiniResponse.builder()
-                    .userId(lookbook.getUser().getUserId())
-                    .images(lookbookImages)
-                    .view(lookbook.getView())
+            List<LookbookImage> lookbookImage = lookbookImageRepository.findAllByLookbook_LookbookId(lookbook.getLookbookId());
+            List<String> ImageUrls=lookbookImage.stream().map(Image-> Image.getImage()).collect(Collectors.toList());
+            Integer likeCnt=likesRepository.countByLookbook(lookbook);
+            User user=lookbook.getUser();
+            LookbookThumbnailResponse build = LookbookThumbnailResponse.builder()
+                    .lookbookId(lookbook.getLookbookId())
                     .createdAt(lookbook.getCreatedAt())
+                    .likeCnt(likeCnt)
+                    .userNickname(user.getNickname())
+                    .userProfileImg(user.getProfileImg())
+                    .images(ImageUrls)
                     .build();
-
-            lookbookResponse.add(response);
+            lookbookResponse.add(build);
         });
 
         return new PagingResponse(
@@ -180,6 +213,7 @@ public class LookbookService {
     @Transactional
     public void updateLookbook(Integer userId, Integer lookbookId, List<MultipartFile> images, LookbookUpdateRequest lookbookUpdateRequest) throws UserNotMatchException {
         Lookbook lookbook = lookbookRepository.findByLookbookId(lookbookId);
+        if(lookbook==null) throw new DataNotFoundException("해당 룩북은 존재하지 않습니다.");
         if (lookbook.getUser().getUserId() != userId) throw new UserNotMatchException("해당 유저에게 룩북 수정 권한이 존재하지 않습니다.");
 
         if (!images.isEmpty()) {
@@ -216,16 +250,17 @@ public class LookbookService {
             tagLookbookRepository.deleteByLookbookLookbookId(lookbookId);
 
             lookbookUpdateRequest.getTags().forEach(tagName -> {
-                Tag tag = Tag.builder()
-                        .name(tagName)
-                        .build();
-                tagRepository.save(tag);
-
+                Tag tag=tagRepository.findByName(tagName);
+                if(tag==null){
+                    tag = Tag.builder()
+                            .name(tagName)
+                            .build();
+                    tagRepository.save(tag);
+                }
                 TagLookbook tagLookbook = TagLookbook.builder()
                         .tag(tag)
                         .lookbook(lookbook)
                         .build();
-
                 tagLookbookRepository.save(tagLookbook);
             });
         }
@@ -234,7 +269,22 @@ public class LookbookService {
     @Transactional
     public void deleteLookbook(Integer userId, Integer lookbookId) {
         Lookbook lookbook = lookbookRepository.findByLookbookId(lookbookId);
+        if(lookbook==null) throw new DataNotFoundException("해당 룩북은 존재하지 않습니다.");
         if (lookbook.getUser().getUserId() != userId) throw new UserNotMatchException("해당 유저에게 룩북 삭제 권한이 존재하지 않습니다.");
+
+        tagLookbookRepository.deleteByLookbookLookbookId(lookbookId); //Tag-Lookbook 삭제
+        commentRepository.deleteByLookbookLookbookId(lookbookId); //comment 삭제
+        bookmarkRepository.deleteByLookbookLookbookId(lookbookId); //bookmark 삭제
+        likesRepository.deleteByLookbookLookbookId(lookbookId); //like 삭제
+        clothesLookbookRepository.deleteAllByLookbook(lookbook); //clothesLookbook삭제
+
+        //lookbookImage 삭제
+        List<LookbookImage> lookbookImages = lookbookImageRepository.findAllByLookbook_LookbookId(lookbookId);
+        lookbookImages.forEach((lookbookImage) -> {
+            s3Uploader.deleteFile(lookbookImage.getImage());
+        });
+        lookbookImageRepository.deleteByLookbook(lookbook);
+
         lookbookRepository.deleteByLookbookId(lookbookId);
     }
 
@@ -243,17 +293,19 @@ public class LookbookService {
         List<LookbookMiniResponse> responseList = new ArrayList<>();
 
         followingList.getFollowingList().forEach(followingId -> {
-            List<Lookbook> lookbooks = lookbookRepository.findByUserUserId(followingId);
-            lookbooks.forEach(lookbook -> {
-                List<LookbookImage> lookbookImages = lookbookImageRepository.findAllByLookbook_LookbookId(lookbook.getLookbookId());
-                LookbookMiniResponse lookbookMiniResponse = LookbookMiniResponse.builder()
-                        .images(lookbookImages)
-                        .createdAt(lookbook.getCreatedAt())
-                        .view(lookbook.getView())
-                        .userId(lookbook.getUser().getUserId())
-                        .build();
-                responseList.add(lookbookMiniResponse);
-            });
+            if(userRepository.existsById(followingId)){
+                List<Lookbook> lookbooks = lookbookRepository.findByUserUserId(followingId);
+                lookbooks.forEach(lookbook -> {
+                    List<LookbookImage> lookbookImages = lookbookImageRepository.findAllByLookbook_LookbookId(lookbook.getLookbookId());
+                    LookbookMiniResponse lookbookMiniResponse = LookbookMiniResponse.builder()
+                            .images(lookbookImages)
+                            .createdAt(lookbook.getCreatedAt())
+                            .view(lookbook.getView())
+                            .userId(lookbook.getUser().getUserId())
+                            .build();
+                    responseList.add(lookbookMiniResponse);
+                });
+            }
         });
         System.out.println(responseList);
         return new PageImpl<>(responseList, pageable, responseList.size());
@@ -280,7 +332,9 @@ public class LookbookService {
     @Transactional
     public void bookmark(Integer userId, Integer lookbookId) {
         Lookbook lookbook = lookbookRepository.findByLookbookId(lookbookId);
-        if (lookbook.getUser().getUserId() != userId) throw new UserNotMatchException("해당 유저에게 북마크 추가 권한이 존재하지 않습니다.");
+        if(lookbook==null) throw new DataNotFoundException("해당 룩북은 존재하지 않습니다.");
+        boolean isBookmarked = bookmarkRepository.existsByLookbookAndUserUserId(lookbook, userId);
+        if(isBookmarked) throw new DataMismatchException("해당 유저는 이미 북마크를 한 상태입니다.");
 
         User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("해당 유저가 존재하지 않습니다."));
         Bookmark bookmark = Bookmark.builder()
@@ -294,7 +348,9 @@ public class LookbookService {
     @Transactional
     public void unbookmark(Integer userId, Integer lookbookId) {
         Lookbook lookbook = lookbookRepository.findByLookbookId(lookbookId);
-        if (lookbook.getUser().getUserId() != userId) throw new UserNotMatchException("해당 유저에게 북마크 해제 권한이 존재하지 않습니다.");
+        if(lookbook==null) throw new DataNotFoundException("해당 룩북은 존재하지 않습니다.");
+        boolean isBookmarked = bookmarkRepository.existsByLookbookAndUserUserId(lookbook, userId);
+        if(!isBookmarked) throw new DataMismatchException("해당 유저는 북마크를 하지 않은 상태입니다.");
 
         bookmarkRepository.deleteByLookbookLookbookId(lookbookId);
     }
@@ -303,6 +359,9 @@ public class LookbookService {
     public void likes(Integer userId, Integer lookbookId) {
         User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("해댱 유저가 존재하지 않습니다."));
         Lookbook lookbook = lookbookRepository.findByLookbookId(lookbookId);
+        if(lookbook==null) throw new DataNotFoundException("해당 룩북은 존재하지 않습니다.");
+        boolean isLiked = likesRepository.existsByLookbookAndUserUserId(lookbook, userId);
+        if(isLiked) throw new DataMismatchException("해당 유저는 이미 좋아요를 한 상태입니다.");
         Likes likes = Likes.builder().lookbook(lookbook).user(user).build();
         likesRepository.save(likes);
     }
@@ -311,7 +370,9 @@ public class LookbookService {
     public void unlikes(Integer userId, Integer lookbookId) {
         User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("해댱 유저가 존재하지 않습니다."));
         Lookbook lookbook = lookbookRepository.findByLookbookId(lookbookId);
-
+        if(lookbook==null) throw new DataNotFoundException("해당 룩북은 존재하지 않습니다.");
+        boolean isLiked = likesRepository.existsByLookbookAndUserUserId(lookbook, userId);
+        if(!isLiked) throw new DataMismatchException("해당 유저는 이미 좋아요를 하지 않은 상태입니다.");
         likesRepository.deleteByLookbookAndUser(lookbook, user);
     }
